% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_votesmart.R
\name{read_votesmart}
\alias{read_votesmart}
\title{Read data tables from VoteSmart}
\usage{
read_votesmart(url)
}
\arguments{
\item{url}{URL of the VoteSmart page you wish to scrape data from}
}
\description{
This function allows you to read data tables from the VoteSmart website.
}
\examples{
read_votesmart()
}
\keyword{VoteSmart}
\keyword{table}
